+++ R/survival_data.R
@@ -1,0 +1,26 @@
+#' @title Survival data as example data for survival_plot function.
+#' @description Survival data as example data for survival_plot function.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include columns (Time, Status, Group), rows (records).
+#' @usage data(survival_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/SurvivalAnalysis/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(survival_data)
+#'
+#' # 3. View example data
+#' survival_data
+#'
+"survival_data"
+
+# survival_data <- data
+# usethis::use_data(survival_data)